SQL DB Character Creation Implemented

+++ gk/gatekeeper.js
@@ -22,39 +22,29 @@ function spanner(chardat, nest) {
     })
 }
 
+
 module.exports = {
+// CREATE CHARACTER DATA //
+    char_create: async function(msg, client) {
+        let welcome = "";
 
-// CREATE PLAYER DATA //
-    create: function(msg, client) {
-        channelName = msg.author.tag + " Character Creation"
-        channelid = msg.guild.channels.create(channelName, {
-            type: 'text',
-            parent: '877407686700982362',
-            permissionOverwrites:[
-            { // everyone
-                id: '876270357236023316',
-                deny:['VIEW_CHANNEL'],
-            },
-            { // moderator
-                id: '876288940351553558',
-                allow: ['VIEW_CHANNEL'],
-            },
-            { // user
-                id: msg.author.id,
-                allow: ['VIEW_CHANNEL']
-            }]
-        });
-
-        member = support.member(msg);
+        let channelID = support.createChannel(msg);
+        let member = support.member(msg);
         member.roles.add(support.roles("birth"));
+        
+        if (!(await player.getPlayerData(msg.author.id))) {
+            await player.addPlayerData(msg.author.id);
+            welcome = "<@" + msg.author.id + ">, Welcome to the World of Desolation!\n"; 
+        }
 
-        channelid.then(function(result) {
-            client.channels.cache.get(result.id).send("<@" + msg.author.id + ">.. Let us begin our journey!\n Type `.add` to create your player.");
+        channelID.then(function(result) {
+            client.channels.cache.get(result.id).send(welcome + "Type `.create` to get started your player.");
         })
     },
 
 // FINISH CREATION //
-    birth: function(msg, client) {
+    // deletes character creation channel
+    join: function(msg, client) {
         member = support.member(msg);
         chan = client.channels.cache.get(msg.channel.id);
         // checks if channel has the appropriate role and user privileges set
@@ -73,7 +63,7 @@ module.exports = {
         let statdat = {};
         let Title = "";
         if (cmd == "player") {
-            statdat = character;
+            statdat = chardata;
             Title = "Character Sheet";
             }
         if (cmd == "focus") {
@@ -144,63 +134,111 @@ module.exports = {
 
 // TESTIES //)
     addchar: async function (msg) {
-       
-        // start character creation
-        await msg.channel.send("Are you ready?");
-        let shortmsg = await support.userprompt(msg);
+        let player_data = await player.getPlayerData(msg.author.id);
+        if (player_data) {
+            let char_count = await player.getCharCount(msg.author.id);
+            if (char_count < 3) {
+                // start character creation
+                await msg.channel.send("Are you ready?");
+                let shortmsg = await support.userprompt(msg);
 
-        if (!shortmsg) {
-            await msg.channel.send("Your response time is too slow.. try paying attention this time.");
-            return;
-        } else
+                if (!shortmsg) {
+                    await msg.channel.send("Your response time is too slow.. try paying attention this time.");
+                    return;
+                } else
+                if (shortmsg[0].toLowerCase() === 'y') {
+                    // instantiate player class
+                    let thisChar = new character.Type();
 
-        if (shortmsg.toLowerCase() === 'yes') {
-            // instantiate player class
-            let thisPlayer = new player.Type();
-            
-            // get player id
-            thisPlayer.player_id = msg.author.id;
+                    // get player id
+                    thisChar.player_id = msg.author.id;
 
-            // get character name
-            thisPlayer.char_name = await thisPlayer.getCharName(msg);
-            if (thisPlayer.char_name) { 
-                // get character race
-                thisPlayer.race = await thisPlayer.getRace(msg);
-                if(thisPlayer.race) {
-                    thisPlayer.level = 1;
-                    thisPlayer.alignment = 'neutral';
-                    thisPlayer.mastery = 'locked';
-                    thisPlayer.discipline = 'none';
-                    thisPlayer.group = 'none';
-                    thisPlayer.quest = ['Genesis'];
-                    db.addNewPlayer(thisPlayer);
-                    msg.channel.send("here we are");
+                    // get character name
+                    thisChar.char_name = await thisChar.getCharName(msg);
+
+                    // get character sex
+                    if (thisChar.char_name) { 
+                        thisChar.sex = await thisChar.getSex(msg);
+
+                        // get character race
+                        if (thisChar.sex) {
+                            thisChar.race = await thisChar.getRace(msg);
+                        } else {
+                            msg.channel.send("Well, we almost got somewhere.\nTry starting over.");
+                        }
+                    } else {
+                        msg.channel.send("You failed the simplest of tasks.\nStart over.");
+                    }
+
+                    // populate beginner attributes
+                    if(thisChar.race) {
+                        thisChar.level = 1;
+                        thisChar.world = "starter";
+                        thisChar.loc = support.locIndex(20, 20);
+                        thisChar.discipline = 'none';
+                        thisChar.mastery = 'locked';
+                        thisChar.alignment = 'neutral';
+                        thisChar.group_name = 'none';
+                        thisChar.quest_current = ['Genesis'];
+                        thisChar.clan = 'none';
+                    }
+
+                // UPDATE DATABASE //
+                    // add new character to database
+                    await db.addNewChar(thisChar);
+                    
+                    // set new character as active character
+                    let activechar = await db.setActiveChar(msg.author.id, thisChar.char_name);
+                    if (activechar){
+                        await msg.channel.send("Well, " + thisChar.char_name + ". It is time for you to `.join` the fight!");
+                    }
+
+                    // update players character list
+                    let char_id = await db.getCharID(msg.author.id, thisChar.char_name);
+                    if (char_id) {
+                        await player.updateCharNames(msg.author.id);
+                    }
+                } else if (shortmsg.toLowerCase() === 'no') {
+                    msg.channel.send("Stop Wasting My Time.");
+                } else {
+                    msg.channel.send("That isn\'t an option.");
                 }
             } else {
-                msg.channel.send("You failed the simplest of tasks.\nStart over.");
+                msg.channel.send("You have too many characters.");
+                // delete this channel
+                return false;
             }
-        } else if (shortmsg.toLowerCase() === 'no') {
-            msg.channel.send("Stop Wasting My Time.");
         } else {
-            msg.channel.send("That isn\'t an option.");
+            msg.channel.send("Why are you here?!");
         }
+
+   },
+
+    log: async function(msg) {
+        let stuff = await player.updateCharNames(msg.author.id); 
+        msg.channel.send("Chars: " + stuff);
+        console.log(stuff);  
     },
 
 // DB HANDLERS //
-    // get player data
+    // add new player to database
     addplayer: async function(msg) {
-        let playerdata = await player.addPlayerData(msg.author.id);
-        if (playerdata) {
+//        let playerdata = await player.addPlayerData(msg.author.id);
+        if (await player.addPlayerData(msg.author.id)) {
             msg.channel.send("Player <@" + support.member(msg) + "> added.");
         } else {
             msg.channel.send("Player already exists. No cheating!");
         }
     },
+
+    // get player info from database
     getplayer: async function(msg) {
         let playerdat = await player.getPlayerData(msg.author.id);
 
         msg.channel.send(playerdat);
     },
+
+    // delete player from database
     delplayer: async function(msg) {
         let playerdata = await player.getPlayerData(msg.author.id);
         if (playerdata) {
@@ -214,43 +252,53 @@ module.exports = {
         }
 
     },
+
+    // get character info
     getchar: async function(msg) {
-        let pd = await db.getChars(msg.author.id);
-        let sections = "";
-
-        for (let chars in pd.rows){
-            embedstring += "```\n";
-            for (let keys in pd.rows[chars]) {
-                if (keys !== 'id' && keys !== 'modified') {
-                    embedstring = embedstring +  keys + ": " + pd.rows[chars][keys] + "\n";
+        let pd = await db.getCharacters(msg.author.id);
+        embedstring = "";
+        
+        if (pd.rows[0]) {
+            for (let chars in pd.rows){
+                embedstring += "```\n";
+                for (let keys in pd.rows[chars]) {
+                    if (keys !== 'id' && keys !== 'modified' &&
+                        keys !== 'player_id' && pd.rows[chars][keys] !== 'locked') {
+                        embedstring = embedstring +  keys + ": " + pd.rows[chars][keys] + "\n";
+                    }
                 }
+                embedstring += "```\n";
             }
-            embedstring += "```\n";
-        }
 
-        embedded = new Discord.MessageEmbed()
-            .setTitle(msg.author.tag + " Character Data:")
-            .setDescription( embedstring );
+            embedded = new Discord.MessageEmbed()
+                .setTitle(msg.author.tag + " Character Data:")
+                .setDescription( embedstring );
  
-        msg.channel.send(embedded);
+            msg.channel.send(embedded);
+        } else {
+            msg.channel.send("You have no character.");
+        }
     },
 
+    // delete character
     delchar: async function(msg, vars) {
-        let pd = await db.getPlayer(msg.author.id);
-
-        for (let chars in pd.rows){
-            if (pd.rows[chars].char_name === vars){
-                msg.channel.send("Are you sure?");
-                let response = await support.userprompt(msg);
-                if (response === 'yes') {
-                    db.deleteCharacter(msg.author.id, vars);
-                    msg.channel.send("RIP " + vars);
+        let pd = await db.getCharacters(msg.author.id);
+        if(pd.rows[0]){
+            for (let chars in pd.rows){
+                if (pd.rows[chars].char_name === vars){
+                    msg.channel.send("Are you sure?");
+                    let response = await support.userprompt(msg);
+                    if (response === 'yes') {
+                        await db.deleteCharacter(msg.author.id, vars);
+                        await player.updateCharNames(msg.author.id);
+                        msg.channel.send("RIP " + vars);
+                    }
                 }
             }
-            else {
-                msg.channel.send("I don't know who you're talking about.")
-            }
+        } else {
+            msg.channel.send("I don't know who you're talking about.")
         }
+        
     }
 }
